+++ .idea/misc.xml
@@ -18,6 +18,7 @@
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_login_exists_frame.xml" value="0.25416666666666665" />
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_main.xml" value="0.23854166666666668" />
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_main_toolbar.xml" value="0.23177083333333334" />
+        <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_mypetapp_chat.xml" value="0.2" />
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_nav_drawer_manager.xml" value="0.4" />
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_nav_drower_pet_keeper.xml" value="0.264" />
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_navigation_drawer.xml" value="0.24507658643326038" />
@@ -28,7 +29,7 @@
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_reliability_form.xml" value="0.264" />
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_reminder.xml" value="0.2" />
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_reports.xml" value="0.13020833333333334" />
-        <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_search_page.xml" value="0.2" />
+        <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_search_page.xml" value="0.176" />
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_second_fragment.xml" value="0.1625" />
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_start_work.xml" value="0.23854166666666668" />
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_technical_support.xml" value="0.264" />
@@ -38,6 +39,7 @@
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_to_do_list.xml" value="0.23854166666666668" />
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_vaccines.xml" value="0.17708333333333334" />
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_vaccines_and_medicines_slide.xml" value="0.17708333333333334" />
+        <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/activity_vaccines_info.xml" value="0.264" />
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/app_bar_base.xml" value="0.17708333333333334" />
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/dialog_search_spinner.xml" value="0.17708333333333334" />
         <entry key="..\:/Users/Ester/Documents/GitHub/MyPetApp/app/src/main/res/layout/dropdown_item.xml" value="0.17708333333333334" />
