+++ app/src/main/AndroidManifest.xml
@@ -69,9 +69,6 @@
             android:name=".tips"
             android:exported="false" />
         <activity
-            android:name=".Reminder"
-            android:exported="false" />
-        <activity
             android:name=".ToDoList"
             android:exported="false" />
         <activity
