--- app/src/main/java/com/example/myproject/Reminder.java
@@ -1,57 +1,0 @@
-package com.example.myproject;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.drawerlayout.widget.DrawerLayout;
-
-import android.app.Activity;
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.MenuItem;
-import android.view.View;
-import android.widget.ImageView;
-
-import com.google.android.material.navigation.NavigationView;
-
-public class Reminder extends AppCompatActivity {
-    DrawerLayout drawerLayout;
-    NavigationView navigation;
-    ImageView MenuIcon,BackIcon;
-    User user = new User();
-    Intent intent;
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_reminder);
-        intent = getIntent();
-        user = (User)intent.getSerializableExtra("user");
-        //assign variable
-        drawerLayout =findViewById(R.id.drawer_layout) ;
-        navigation = findViewById(R.id.NavigationView);
-        navigation.setNavigationItemSelectedListener(new NavigationView.OnNavigationItemSelectedListener() {
-            @Override
-            public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-                new PetKeeperNavigation(Reminder.this,item.getItemId(),user);
-                return false;
-            }
-        });
-        MenuIcon = findViewById(R.id.MenuItem);
-        BackIcon = findViewById(R.id.BackItem);
-        MenuIcon.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                drawerLayout.open();
-            }
-        });
-        BackIcon.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                intent = new Intent(Reminder.this, navigation_drawer.class);
-                intent.putExtra("user", user);
-                startActivity(intent);
-                finish();
-            }
-        });
-    }
-
-}
